[libc++] Quick fix to remove a narrowing conversion from the in_out_result test.
This needs a proper solution in a follow-up. The issue is that the
Standard defines conversions between `in_out_result` classes with
different template types as just `return {in, out};`. Because the
expression uses list initialization, it will fail to compile if the
conversion happens to be narrowing -- which is probably unintended.

Surprisingly, this error wasn't caught by the CI.

Differential Revision: https://reviews.llvm.org/D117089

+++ libcxx/test/std/algorithms/algorithms.results/in_out_result.pass.cpp
@@ -53,8 +53,9 @@ int main(int, char**) {
   // Conversion, fundamental types.
   {
     std::ranges::in_out_result<int, bool> x = {2, false};
-    std::ranges::in_out_result<double, char> y = x;
-    assert(y.in == 2.0);
+    // FIXME(varconst): try a narrowing conversion.
+    std::ranges::in_out_result<long, char> y = x;
+    assert(y.in == 2);
     assert(y.out == '\0');
   }
 
@@ -138,4 +139,3 @@ int main(int, char**) {
 
   return 0;
 }
-
